requirements.txt
mido[ports-rtmidi]
requests
pyyaml
